+++ src/import_export_csv/templates/import_export_csv/locatie-import.html
@@ -32,7 +32,7 @@ Import
         {% endfor %}
     </fieldset>
     <div class="btn-container">
-        <button class="btn btn-primair" type="submit" name="_save" value="Import vouchers" formaction="{% url 'locations_urls:location-import' %}">
+        <button class="btn btn-primair" type="submit" name="_save" value="Import vouchers" formaction="{% url 'import_export_urls:locatie-import' %}">
             Importeren
         </button>
     </div>
