+++ src/import_export_csv/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from import_export_csv.views import LocatieImportView
+
+urlpatterns = [
+    path("import", view=LocatieImportView.as_view(), name="locatie-import"),
+]
