+++ src/locations/templates/locations/location-admin.html
@@ -62,17 +62,6 @@
         | Voor het toevoegen, wijzigen van externe gegevensbronnen.
     </li>
     <li>
-        <a href="{% url 'locations_urls:location-import' %}">
-            <span class="icon">
-                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32" aria-hidden="true" focusable="false">
-                    <path fill-rule="evenodd" d="m9.757 32-2.9-2.91L19.937 16 6.857 2.91 9.757 0l16 16z"></path>
-                </svg>
-            </span>
-            Data importeren
-        </a>
-        | Voor het importeren van nieuwe en gewijzigde locaties middels een CSV bestand.
-    </li>
-    <li>
         <a href="{% url 'locations_urls:location-export' %}">
             <span class="icon">
                 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32" aria-hidden="true" focusable="false">
