+++ src/locations/urls.py
@@ -9,7 +9,6 @@ from locations.views import (
     LocationCreateView,
     LocationDetailView,
     LocationExportView,
-    LocationImportView,
     LocationListView,
     LocationLogView,
     LocationPropertyCreateView,
@@ -34,7 +33,6 @@ urlpatterns = [
     path("toevoegen", view=LocationCreateView.as_view(), name="location-create"),
     path("beheer", view=LocationAdminView.as_view(), name="location-admin"),
     path("export", view=LocationExportView.as_view(), name="location-export"),
-    path("import", LocationImportView.as_view(), name="location-import"),
     path("log/", view=LocationLogView.as_view(), name="location-log"),
     path("log/locatie/<int:pandcode>", view=LocationLogView.as_view(), name="location-detail-log"),
     path("locatie-eigenschappen/", view=LocationPropertyListView.as_view(), name="locationproperty-list"),
